applications: sdp: mspi: fix for data longer than 256 bytes
Changed type uint8_t to uint32_t to be able to
send data longer than 256 bytes.

+++ applications/sdp/mspi/src/hrt/hrt.c
@@ -314,7 +314,7 @@ void hrt_read(volatile hrt_xfer_t *hrt_xfer_params)
 	hrt_tx_rx(&hrt_xfer_params->xfer_data[HRT_FE_ADDRESS], hrt_xfer_params->bus_widths.address,
 		  false, hrt_xfer_params->counter_value, CNT1_INIT_VALUE);
 
-	for (uint8_t i = 0; i < hrt_xfer_params->xfer_data[HRT_FE_DATA].word_count; i++) {
+	for (uint32_t i = 0; i < hrt_xfer_params->xfer_data[HRT_FE_DATA].word_count; i++) {
 		hrt_xfer_params->xfer_data[HRT_FE_DATA].data[i] =
 			nrf_vpr_csr_vio_in_buffered_reversed_byte_get() >> INPUT_SHIFT_COUNT;
 	}
